+++ addons/mail/tests/discuss/test_message_controller.py
@@ -3,8 +3,10 @@
 import json
 
 import odoo
+from odoo.tests import tagged, users
 from odoo.tools import mute_logger
-from odoo.addons.base.tests.common import HttpCaseWithUserDemo
+from odoo.addons.base.tests.common import HttpCase, HttpCaseWithUserDemo
+from odoo.addons.mail.tests.common import MailCommon, mail_new_test_user
 from odoo.http import STATIC_CACHE_LONG
 from odoo import Command, fields
 
@@ -335,3 +337,41 @@ class TestMessageController(HttpCaseWithUserDemo):
             url=f"/web/image?field=avatar_128&id={self.guest.id}&model=mail.guest"
         )
         self.assertIn("no-cache", res.headers["Cache-Control"])
+
+
+@tagged("mail_message")
+class TestMessageLinks(MailCommon, HttpCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+
+        cls.user_employee_1 = mail_new_test_user(cls.env, login='tao1', groups='base.group_user', name='Tao Lee')
+        cls.public_channel = cls.env['discuss.channel'].channel_create(name='Public Channel1', group_id=None)
+        cls.private_group = cls.env['discuss.channel'].create_group(partners_to=cls.user_employee_1.partner_id.ids, name="Group")
+
+    @users('employee')
+    def test_message_link_by_employee(self):
+        channel_message = self.public_channel.message_post(body='Public Channel Message', message_type='comment')
+        private_message_id = self.private_group.with_user(self.user_employee_1).message_post(
+            body='Private Message',
+            message_type='comment',
+        ).id
+        self.authenticate('employee', 'employee')
+        with self.subTest(channel_message=channel_message):
+            expected_url = self.base_url() + f'/odoo/action-mail.action_discuss?active_id={channel_message.res_id}&highlight_message_id={channel_message.id}'
+            res = self.url_open(f'/mail/message/{channel_message.id}')
+            self.assertEqual(res.url, expected_url)
+        with self.subTest(private_message_id=private_message_id):
+            res = self.url_open(f'/mail/message/{private_message_id}')
+            self.assertEqual(res.status_code, 401)
+
+    @users('employee')
+    def test_message_link_by_public(self):
+        message = self.public_channel.message_post(
+            body='Public Channel Message',
+            message_type='comment',
+            subtype_xmlid='mail.mt_comment'
+        )
+        res = self.url_open(f'/mail/message/{message.id}')
+        self.assertEqual(res.status_code, 200)
